Cleaned up Stew's API code, added more methods to Device class.
Not sure how to handle this.

One one hand, if the device is passed through query parameters it makes for
cleaner code by using an __init__ method to initialize the td.Device class.

It also allows for referencing the device by its index, or by its complex
name in query params.

On the other hand, REST API's look a lot more proper if they address the
device in the URL path like so: /device/\d+/method.

Open to suggestions and discussions.